Fix binomial variance correction
The binomial variance we calculate (curBinoVar) is the *expected*
squared difference to the expected value. In hindsight it obviously
doesn't make any sense to correct for the binomial variance at the level
of individual axons.

Some axons will have a larger squared difference than the variance, some
axons will have a lower squared difference. But in the latter case,
where the binomial variance is larger than the squared difference, we
cannot throw away the surplus binomial variance. We have to use it for
the former cases, where the squared difference is larger than the
expected squared difference.

In practice, this means that the `max(... - curBinoVar, 0)` was just
plain wrong. This is fixed now and essentially undoes the "bad" commit:

commit 051b6bc30d863375a7c2ace5df203d242cf84ace
Author: Alessandro Motta <alessandro.motta@brain.mpg.de>
Date:   Sat Aug 11 20:10:05 2018 +0200

    Subtract per-axon binomial variance, instead of constant mean value

Furthermore, I've realized that not only the squared residuals, but also
the squared difference to the mean must be corrected for binomial
variance. This ensures that R² goes towards zero for the case where the
mean value is predicted.